STOS policy engine unit tests
